Updated comments
- Updated comments
- Moved unit test helper functions into single file
