+++ src/app/component/shared/menu/external-link/external-link.component.spec.ts
@@ -1,9 +1,9 @@
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { RouterTestingModule } from "@angular/router/testing";
-import { assertIcon, assertTooltip } from "src/app/helpers/tests/helpers";
 import { MenuLink } from "src/app/interfaces/menusInterfaces";
 import { AppConfigService } from "src/app/services/app-config/app-config.service";
 import { testAppInitializer } from "src/app/test.helper";
+import { assertIcon, assertTooltip } from "src/testHelpers";
 import { SharedModule } from "../../shared.module";
 import { MenuExternalLinkComponent } from "./external-link.component";
