+++ src/app/component/shared/menu/internal-link/internal-link.component.spec.ts
@@ -1,12 +1,8 @@
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { RouterTestingModule } from "@angular/router/testing";
-import {
-  assertIcon,
-  assertRoute,
-  assertTooltip
-} from "src/app/helpers/tests/helpers";
 import { MenuRoute } from "src/app/interfaces/menusInterfaces";
 import { StrongRoute } from "src/app/interfaces/strongRoute";
+import { assertIcon, assertRoute, assertTooltip } from "src/testHelpers";
 import { SharedModule } from "../../shared.module";
 import { MenuInternalLinkComponent } from "./internal-link.component";
