+++ src/app/directives/directives.module.ts
@@ -3,6 +3,9 @@ import { DatatableDirective } from "./datatable/datatable.directive";
 
 const directives = [DatatableDirective];
 
+/**
+ * App Shared Directives
+ */
 @NgModule({
   declarations: directives,
   exports: directives
