+++ src/app/services/baw-api/resolver-common.spec.ts
@@ -1,0 +1,6 @@
+// TODO Write unit tests
+xdescribe("API Resolvers", () => {
+  describe("Resolvers", () => {});
+  describe("ListResolver", () => {});
+  describe("ShowResolver", () => {});
+});
